Make executable and library versions of the worker use the same exact code
