+++ worker/include/lib.hpp
@@ -1,0 +1,10 @@
+extern "C" int run_worker(
+    int argc,
+    char* argv[],
+    const char* version,
+    bool processMode,
+    int consumerChannelFd,
+    int producerChannelFd,
+    int payloadConsumeChannelFd,
+    int payloadProduceChannelFd
+);
