+++ worker/mediasoup-worker.gyp
@@ -16,6 +16,7 @@
     'sources':
     [
       # C++ source files.
+      'src/lib.cpp',
       'src/DepLibSRTP.cpp',
       'src/DepLibUV.cpp',
       'src/DepLibWebRTC.cpp',
@@ -127,6 +128,7 @@
       'src/RTC/RTCP/XrDelaySinceLastRr.cpp',
       'src/RTC/RTCP/XrReceiverReferenceTime.cpp',
       # C++ include files.
+      'include/lib.hpp',
       'include/DepLibSRTP.hpp',
       'include/DepLibUV.hpp',
       'include/DepLibWebRTC.hpp',
@@ -334,12 +336,7 @@
     },
     {
       'target_name': 'libmediasoup-worker',
-      'type': 'static_library',
-      'sources':
-      [
-        # C++ source files.
-        'src/lib.cpp'
-      ]
+      'type': 'static_library'
     },
     {
       'target_name': 'mediasoup-worker-test',
